% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{blank,PlateData-method}
\alias{blank,PlateData-method}
\title{Accessors for the 'data' element of an PlateData object.}
\usage{
\S4method{blank}{PlateData}(x)
}
\description{
The \code{blank} slot in an PlateData object holds
a data.frame containing blank measurements extracted from combined(object)
}
\examples{
blank(object)
}
\author{
Oliver Dietrich
}
