% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_layout_and_data}
\alias{combine_layout_and_data}
\title{Combine layout and data}
\usage{
combine_layout_and_data(data = NULL, layout = NULL, match_key = "well")
}
\arguments{
\item{data}{Data.frame with data}

\item{layout}{Data.frame with layout}

\item{match_key}{Column name for matching both data.frames (default: well)}
}
\description{
Combine data.frames with data and layout
}
