% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{combined,PlateData-method}
\alias{combined,PlateData-method}
\title{Accessors for the 'combined' element of an PlateData object.}
\usage{
\S4method{combined}{PlateData}(x)
}
\description{
The \code{combined} slot in an PlateData object holds
a data.frame containing data combined from raw and layout
}
\examples{
combined(object)
}
\author{
Oliver Dietrich
}
