% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{control,PlateData-method}
\alias{control,PlateData-method}
\title{Accessors for the 'control' element of an PlateData object.}
\usage{
\S4method{control}{PlateData}(x)
}
\description{
The \code{control} slot in an PlateData object holds
a data.frame containing blank measurements extracted from combined(object)
}
\examples{
control(object)
}
\author{
Oliver Dietrich
}
