% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{CreatePlateData}
\alias{CreatePlateData}
\title{Create a \code{PlateData} object}
\usage{
CreatePlateData(
  layout,
  raw,
  value,
  time,
  predictors,
  treatment,
  blanks = NULL,
  controls = NULL,
  ...
)
}
\arguments{
\item{layout}{Data.frame containing columns 'plate' and 'well' alongside metadata}

\item{value}{Measurement values}

\item{time}{Name of the column that defines longitudinal measurements (e.g. 'Hours' for time series).}

\item{predictors}{Column(s) that predict treatment effect on value (e.g. Bacterial strain AND Plasmid)}

\item{treatment}{Column with treatment variable (e.g. Phage)}

\item{blanks}{Boolean vector indicating layout rows serving as blank}

\item{controls}{Boolean vector indicating layout rows serving as controls}

\item{data}{Data.frame containing columns \code{plate} and \code{well} alongside \code{value} and \code{time}}
}
\value{
A \code{\link{PlateData}} object
}
\description{
Create a \code{PlateData} object from layout and data
}
