% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{data,PlateData-method}
\alias{data,PlateData-method}
\title{Accessors for the 'data' element of an PlateData object.}
\usage{
\S4method{data}{PlateData}(x)
}
\description{
The \code{data} slot in an PlateData object holds
a data.frame containing measurements attributable to
individual wells. Can contain multiple rows per well.
}
\examples{
data(object)
}
\author{
Oliver Dietrich
}
