% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_largest_plate_type}
\alias{detect_largest_plate_type}
\title{Detect largest plate type}
\usage{
detect_largest_plate_type(object)
}
\arguments{
\item{object}{Data.frame with columns 'plate', 'row', and 'col'}
}
\value{
Character vector with only one plate type
}
\description{
Detect largest plate type
}
