% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_plate_type}
\alias{detect_plate_type}
\title{Detect plate type}
\usage{
detect_plate_type(object, plate_types = plate_types_default)
}
\arguments{
\item{object}{Data.frame with columns 'plate', 'row', and 'col'}

\item{plate_types}{Plate types available for dummyPlate

Stores the 'last_well' for a number of commonly used plate types,
indicating the bottom right corner of a plate to distinguish
plate dimensions.}
}
\value{
Named vector of plate type per plate
}
\description{
Detect the plate type (e.g. 6-, 24- or 96-well) from the well names. Assigns
one type per plate. Based on sorting the row and column indices and looking
for the closest plate type.
}
