% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummyPlate}
\alias{dummyPlate}
\title{Create dummy plate}
\usage{
dummyPlate(type = NULL, last_well = NULL)
}
\arguments{
\item{type}{Name of layout (e.g. 6-well, also for 12, 24, 48, 96)}

\item{last_well}{Name of the last well on the plate (e.g. H12)}
}
\value{
Data.frame with layout of microtiter plate
}
\description{
Creates a layout sheet of an empty microtiter well. The size of the plate
is determined either by specifying a type (e.g. 96-well) or by the bottom right corner (e.g. H12 of a 96-well plate).
}
