% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_elecsys}
\alias{import_elecsys}
\title{Import data from Elecsys}
\usage{
import_elecsys(
  file_path = NULL,
  run_name = "unspecified",
  sep = ";",
  plate_well = "PID"
)
}
\arguments{
\item{file_path}{Path to file}

\item{run_name}{Name of run}

\item{sep}{Separator used in CSV}

\item{plate_well}{Column storing plate and well, separated by underscore (_)}
}
\description{
Import data from Elecsys
}
