% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_plateLayout}
\alias{import_plateLayout}
\title{Import plate layout}
\usage{
import_plateLayout(file_path = NULL, plate_name = "unspecified")
}
\arguments{
\item{file_path}{Location of file}

\item{plate_name}{Name of the plate the layout is associated with}

\item{well_key}{Column indicating wells in microtiter plate}
}
\description{
Import plate layout from excel file
}
