% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{layout,PlateData-method}
\alias{layout,PlateData-method}
\title{Accessors for the 'layout' element of an PlateData object.}
\usage{
\S4method{layout}{PlateData}(x)
}
\description{
The \code{layout} slot in an PlateData object holds
a data.frame containing information attributable to
individual wells. Contains one row per unique well.
}
\examples{
layout(pd)
}
\author{
Oliver Dietrich
}
