% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meannorm}
\alias{meannorm}
\title{Mean normalization}
\usage{
meannorm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Based on the formula {\displaystyle x'={\frac {x-{\bar {x}}}{{\text{max}}(x)-{\text{min}}(x)}}}
}
