% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_data_to_layout}
\alias{merge_data_to_layout}
\title{Combine layout and data}
\usage{
merge_data_to_layout(object = NULL)
}
\arguments{
\item{data}{Data.frame with data}

\item{layout}{Data.frame with layout}

\item{match_key}{Column name for matching both data.frames (default: well)}
}
\description{
Combine data.frames with data and layout
}
