% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_well_to_row_col_indices}
\alias{mutate_well_to_row_col_indices}
\title{Mutate well to row and column indices}
\usage{
mutate_well_to_row_col_indices(object)
}
\arguments{
\item{object}{data.frame with column well}
}
\description{
Mutate well to row and column indices
}
