% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate_types}
\alias{plate_types}
\title{Plate types available for dummyPlate}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\usage{
plate_types
}
\description{
Stores the 'last_well' for a number of commonly used plate types,
indicating the bottom right corner of a plate to distinguish
plate dimensions.
}
\keyword{datasets}
