% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{data}
\name{plate_types_default}
\alias{plate_types_default}
\title{Default plate types}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\usage{
plate_types_default
}
\description{
Default plate types
}
\keyword{datasets}
