% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PlateData-class}
\alias{PlateData-class}
\title{The PlateData Class}
\description{
The PlateData object contains measurements that are associated to a specific well in a microtiter plate.
Data from multiple plates can be stored together
}
\section{Slots}{

\describe{
\item{\code{data}}{Data.frame associating measurements and metadata with well and plate indices}

\item{\code{layout}}{Data.frame containing the plate layout}

\item{\code{key}}{Character vector indicating the column to use as key between layout and data}

\item{\code{type}}{Character vector indicating plate type}

\item{\code{misc}}{A list of miscellaneous information}
}}

