% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot curves}
\usage{
plot_curves(
  object,
  slot = "data",
  type = "raw",
  theme.size = 20,
  legend.position = "top"
)
}
\arguments{
\item{object}{PlateData object}

\item{slot}{Slot to get data from (data, summary)}

\item{type}{Whether to show 'raw', 'corrected', or 'normalised' data. Corrected data has blank subtracted. Normalised data is divided by control.}

\item{legend.position}{}
}
\description{
Visualize PlateData as curves, i.e. values by time.
...
}
