% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{plot_plateLayout}
\alias{plot_plateLayout}
\title{Plot the plate layout}
\usage{
plot_plateLayout(
  object = NULL,
  color = NULL,
  row.name = "row",
  col.name = "col",
  pl.title = NULL
)
}
\arguments{
\item{object}{Data.frame containing layout}

\item{color}{Column used to color wells}

\item{row.name}{Column indicating plate rows}

\item{col.name}{Column indicating plate columns}
}
\description{
Plot the plate layout
}
