% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_series}
\alias{plot_series}
\title{Plot series data}
\usage{
plot_series(
  object = NULL,
  x = NULL,
  y = NULL,
  col = NULL,
  group = key(object),
  facet = "plate",
  facet_scales = "fixed",
  facet_rows = NULL,
  theme_size = 20,
  pl_title = "Series",
  y_transform = NULL,
  x_transform = NULL,
  xlab = x,
  ylab = y,
  pl.barwidth = 1,
  pl.barheight = 10,
  color.pal = "RdYlBu",
  color.pal.dir = -1,
  color.values = NULL,
  ...
)
}
\arguments{
\item{object}{PlateData object}

\item{x}{Column used for y axis}

\item{y}{Column used for x axis}

\item{col}{Column name to color measurements by}

\item{group}{Column used to group measurements}

\item{facet}{Column used for faceting}
}
\description{
Visualize numeric measurements stored in a PlateData object.
Both x and y axes must be numeric. Lines will be drawn through grouped data points.
Can be faceted by categorical vectors.
}
