% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{plot_timecourse}
\alias{plot_timecourse}
\title{Plot time course data}
\usage{
plot_timecourse(
  object = NULL,
  x = "time",
  y = "count",
  col = "phage",
  group = "well",
  facet_x = "mode",
  facet_y = "host_number",
  facet_scales = "free",
  theme_size = 20,
  pl_title = "Timecourse",
  y_transform = NULL,
  x_transform = NULL,
  xlab = "Time in hours",
  ylab = NULL
)
}
\arguments{
\item{object}{Data.frame}

\item{x}{Column used for y axis}

\item{y}{Column used for x axis}

\item{col}{Column name to color measurements by}

\item{group}{Column used to group measurements}
}
\description{
Visualize measurements across time
}
