% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{raw,PlateData-method}
\alias{raw,PlateData-method}
\title{Accessors for the 'raw' element of an PlateData object.}
\usage{
\S4method{raw}{PlateData}(x)
}
\description{
The \code{raw} slot in an PlateData object holds
a data.frame containing raw data imported from device-specific
output files formatted by import_* functions
}
\examples{
raw(object)
}
\author{
Oliver Dietrich
}
