% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_plateLayout}
\alias{read_plateLayout}
\title{Read plate layout}
\usage{
read_plateLayout(file_path = NULL, well_key = "well")
}
\arguments{
\item{file_path}{Location of file}

\item{well_key}{Column indicating wells in microtiter plate}
}
\description{
Read plate layout
}
