% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_tekanSpark}
\alias{read_tekanSpark}
\title{Read data from Tekan Spark}
\usage{
read_tekanSpark(file_path = NULL)
}
\arguments{
\item{file_path}{Location of file (must be xls or xlsx)}
}
\value{
Tidy data frame with six columns (cycle, time, temp, well, count, mode)
}
\description{
Read data from Tekan Spark
}
