% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{summary,PlateData-method}
\alias{summary,PlateData-method}
\title{Accessors for the 'summary' element of an PlateData object.}
\usage{
\S4method{summary}{PlateData}(x)
}
\description{
The \code{summary} slot in an PlateData object holds
a data.frame containing summaries from data not attributable to wells
but grouped by 'Plate' and 'Hours'
}
\examples{
summary(object)
}
\author{
Oliver Dietrich
}
