% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{CreatePlateData}
\alias{CreatePlateData}
\title{Create a \code{PlateData} object}
\usage{
CreatePlateData(layout, data = data.frame(), misc = list(), index = NULL, ...)
}
\arguments{
\item{layout}{Data.frame containing columns 'plate' and 'well' alongside metadata}

\item{data}{Data.frame containing columns \code{plate} and \code{well} alongside measurements}

\item{misc}{List of miscellaneous information}
}
\value{
A \code{\link{PlateData}} object
}
\description{
Create a \code{PlateData} object from layout
}
