% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateData_methods.R
\name{CreatePlateData}
\alias{CreatePlateData}
\title{Create a \code{PlateData} object}
\usage{
CreatePlateData(
  layout,
  data = NULL,
  misc = list(),
  key = "key",
  remove_unregistered_wells = FALSE,
  add_unregistered_wells = FALSE,
  ...
)
}
\arguments{
\item{layout}{Data.frame containing columns 'plate' and 'well' alongside metadata}

\item{data}{Data.frame containing columns \code{plate} and \code{well} alongside measurements}

\item{misc}{List of miscellaneous information}

\item{key}{Key to match data.frames between layout and data}

\item{remove_unregistered_wells}{Whether to remove wells from data that do not appear in layout}

\item{add_unregistered_wells}{Whether to add wells that appear in data to layout}
}
\value{
A \code{\link{PlateData}} object
}
\description{
Create a \code{PlateData} object from layout
}
