% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_plate_type}
\alias{detect_plate_type}
\title{Detect plate type}
\usage{
detect_plate_type(object)
}
\arguments{
\item{object}{Data.frame with columns 'plate', 'row', and 'col'}
}
\value{
Named vector of plate type per plate
}
\description{
Detect the plate type (e.g. 6-, 24- or 96-well) from the well names. Assigns
one type per plate. Based on sorting the row and column indices and looking
for the closest plate type.
}
