% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_plateLayout}
\alias{plot_plateLayout}
\title{Plot the plate layout}
\usage{
plot_plateLayout(
  object = NULL,
  color = NULL,
  facet_rows = 1,
  pt.size = 8,
  theme.size = 20,
  pl.title = "Plate layout",
  xlab = NULL,
  ylab = NULL,
  pl.barwidth = 1,
  pl.barheight = 10,
  color.pal = NULL,
  color.pal.dir = 1,
  color.values = NULL,
  color.pal.cols = NULL
)
}
\arguments{
\item{object}{PlateData object}

\item{color}{Column used to color wells}

\item{facet_rows}{Number of rows to use for facet_wrap}
}
\description{
Plot the plate layout
}
